iSOUPTreeRegressor nodes merged and streamlined
